test: Adds async tests
These test async execute, resolve, subscription resolve, and subscription functionality.
